requirements.txt
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
cryptography==40.0.2
cssselect==1.2.0
Cython==0.29.36
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.103.0
filelock==3.12.0
fsspec==2023.6.0
h11==0.14.0
hdbscan==0.8.33
huggingface-hub==0.16.4
hyperlink==21.0.0
idna==3.4
itemadapter==0.8.0
itemloaders==1.0.6
Jinja2==3.1.2
joblib==1.3.2
llvmlite==0.40.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numba==0.57.1
numpy==1.24.4
packaging==23.1
pandas==2.0.3
parsel==1.8.1
Pillow==10.0.0
Protego==0.2.1
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.5.0
pynndescent==0.5.10
pyOpenSSL==23.1.1
pyre-extensions==0.0.29
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.28.2
requests-file==1.5.1
safetensors==0.3.3
scikit-learn==1.2.1
scipy==1.11.2
Scrapy==2.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
tldextract==3.4.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.32.1
Twisted==22.10.0
typing_extensions==4.7.1
tzdata==2023.3
umap-learn==0.5.3
urllib3==1.26.16
uvicorn==0.23.2
xformers==0.0.20
